+++ src/main/java/com/ycloud/client/ApiClient.java
@@ -131,7 +131,7 @@ public class ApiClient {
         json = new JSON();
 
         // Set default User-Agent.
-        setUserAgent("ycloud-sdk-java/1.6.1");
+        setUserAgent("ycloud-sdk-java/1.7.0");
 
         authentications = new HashMap<String, Authentication>();
     }