+++ src/main/java/com/ycloud/client/model/WhatsappMessage.java
@@ -227,7 +227,7 @@ public class WhatsappMessage {
 
   public static final String SERIALIZED_NAME_TOTAL_PRICE = "totalPrice";
   /**
-   * Total price of this message.
+   * Total price of this message. **Note: It&#39;s only an estimated price when the &#x60;status&#x60; is &#x60;accepted&#x60; or &#x60;sent&#x60;. It becomes the final price after the message is delivered, i.e., the &#x60;status&#x60; is &#x60;delivered&#x60; or &#x60;read&#x60;.**
   **/
   @SerializedName(SERIALIZED_NAME_TOTAL_PRICE)
   private Double totalPrice;
@@ -912,7 +912,7 @@ public class WhatsappMessage {
   }
 
    /**
-   * Total price of this message.
+   * Total price of this message. **Note: It&#39;s only an estimated price when the &#x60;status&#x60; is &#x60;accepted&#x60; or &#x60;sent&#x60;. It becomes the final price after the message is delivered, i.e., the &#x60;status&#x60; is &#x60;delivered&#x60; or &#x60;read&#x60;.**
    * @return totalPrice
   **/
   @javax.annotation.Nullable
@@ -923,9 +923,9 @@ public class WhatsappMessage {
 
 
   /**
-   * Total price of this message.
+   * Total price of this message. **Note: It&#39;s only an estimated price when the &#x60;status&#x60; is &#x60;accepted&#x60; or &#x60;sent&#x60;. It becomes the final price after the message is delivered, i.e., the &#x60;status&#x60; is &#x60;delivered&#x60; or &#x60;read&#x60;.**
    **/
-  @ApiModelProperty(example = "0.05", value = "Total price of this message.")
+  @ApiModelProperty(example = "0.05", value = "Total price of this message. **Note: It's only an estimated price when the `status` is `accepted` or `sent`. It becomes the final price after the message is delivered, i.e., the `status` is `delivered` or `read`.**")
   public void setTotalPrice(Double totalPrice) {
     this.totalPrice = totalPrice;
   }